add functionality to get user by userId.

+++ src/main/java/org/ucsc/railboostbackend/models/Role.java
@@ -10,7 +10,7 @@ public class Role {
         this.roleId = roleId;
     }
 
-    public Role(short roleId, String role) {
+    public Role(int roleId, String role) {
         this.roleId = roleId;
         this.role = role;
     }
